requirements.txt
gymnasium[box2d]>=1.0.0
stable-baselines3>=2.0.0
tensorboard>=2.14.0
torch>=2.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
imageio[ffmpeg]>=2.31.0
Pillow>=10.0.0
pytest>=7.0.0
